Implement per-room subscription
