+++ hooks/useSocketClient.tsx
@@ -1,15 +1,15 @@
 import type { ReactNode } from 'react';
-import { useRef, useEffect, createContext, useContext, useMemo, useState } from 'react';
+import { createContext, useContext, useEffect, useMemo, useRef, useState } from 'react';
 import { BehaviorSubject } from 'rxjs';
 import type { Socket } from 'socket.io-client';
 import io from 'socket.io-client';
 
 import { socketsHost } from 'config/constants';
 import log from 'lib/log';
-import { typedKeys } from 'lib/utilities/objects';
-import type { WebsocketEvent, WebsocketMessage } from 'lib/websockets/broadcaster';
-import { WebsocketEvents } from 'lib/websockets/broadcaster';
+import type { WebsocketEvent, WebsocketMessage } from 'lib/websockets/interfaces';
+import { WebsocketEvents } from 'lib/websockets/interfaces';
 
+import { useCurrentSpace } from './useCurrentSpace';
 import { useUser } from './useUser';
 
 type LoggedMessage = { type: string, payload: any }
@@ -19,9 +19,6 @@ type EventFeed = {
 }
 
 type IContext = {
-  lastPong: string | null;
-  isConnected: boolean;
-  sendPing: () => void;
   sendMessage: (message: any) => void;
   // Testing purposes
   messageLog: LoggedMessage[];
@@ -30,9 +27,6 @@ type IContext = {
 }
 
 const WebSocketClientContext = createContext<Readonly<IContext>>({
-  lastPong: null,
-  isConnected: false,
-  sendPing: () => null,
   sendMessage: () => null,
   // Development only
   messageLog: [],
@@ -40,15 +34,16 @@ const WebSocketClientContext = createContext<Readonly<IContext>>({
   eventFeed: {} as any
 });
 
-let socket: Socket;
+let socket: Socket<{ message: (message: WebsocketMessage) => void }>;
 
 export function WebSocketClientProvider ({ children }: { children: ReactNode }) {
 
-  const [isConnected, setIsConnected] = useState(socket?.connected ?? false);
-  const [lastPong, setLastPong] = useState<string | null>(null);
   const [messageLog, setMessageLog] = useState<LoggedMessage[]>([]);
-  const { current: eventFeed } = useRef<EventFeed>(typedKeys(WebsocketEvents).reduce((acc, key) => {
-    acc[key as WebsocketEvent] = new BehaviorSubject<WebsocketMessage | null>(null) as any;
+
+  const [space] = useCurrentSpace();
+
+  const { current: eventFeed } = useRef<EventFeed>(WebsocketEvents.reduce((acc, key) => {
+    acc[key] = new BehaviorSubject<WebsocketMessage | null>(null) as any;
     return acc;
   }, {} as EventFeed));
 
@@ -56,7 +51,7 @@ export function WebSocketClientProvider ({ children }: { children: ReactNode }) 
 
   useEffect(() => {
 
-    if (user) {
+    if (space) {
       connect();
     }
 
@@ -64,7 +59,7 @@ export function WebSocketClientProvider ({ children }: { children: ReactNode }) 
       socket?.disconnect();
       socket?.off();
     };
-  }, [user]);
+  }, [space, user]);
 
   function pushToMessageLog (message: LoggedMessage) {
     if (process.env.NODE_ENV === 'development') {
@@ -87,12 +82,20 @@ export function WebSocketClientProvider ({ children }: { children: ReactNode }) 
     socket.on('connect', () => {
       log.info('Socket client connected');
       pushToMessageLog({ type: 'connect', payload: 'Client connected' });
+      socket.emit('message', {
+        type: 'subscribe',
+        payload: {
+          // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+          spaceId: space!.id
+        }
+      });
     });
 
     socket.on('message', (message: WebsocketMessage) => {
       const isValidMessage = !!message && eventFeed[message.type] !== undefined;
 
       if (isValidMessage) {
+        // Key part when we relay messages from the server to consumers
         (eventFeed[message.type] as BehaviorSubject<WebsocketMessage>).next(message);
         pushToMessageLog(message);
       }
@@ -109,10 +112,6 @@ export function WebSocketClientProvider ({ children }: { children: ReactNode }) 
 
   }
 
-  function sendPing () {
-    socket.emit('ping');
-  }
-
   function sendMessage (message: WebsocketMessage) {
     pushToMessageLog(message);
     socket.emit('message', message);
@@ -123,14 +122,11 @@ export function WebSocketClientProvider ({ children }: { children: ReactNode }) 
   }
 
   const value: IContext = useMemo(() => ({
-    isConnected,
-    lastPong,
-    sendPing,
     sendMessage,
     messageLog,
     clearLog,
     eventFeed
-  }), [isConnected, messageLog]);
+  }), [messageLog]);
 
   return (
     <WebSocketClientContext.Provider value={value}>
