+++ lib/websockets/relay.ts
@@ -10,5 +10,5 @@ export const relay = global.relay ?? new WebsocketBroadcaster();
 
 // remember this instance of prisma in development to avoid too many clients
 if (process.env.NODE_ENV === 'development') {
-  global.prisma = prisma;
+  global.relay = relay;
 }
